ConstructorInterceptor class nuked
