+++ src/main/java/sa/com/cloudsolutions/antikythera/evaluator/MethodInterceptor.java
@@ -1,5 +1,6 @@
 package sa.com.cloudsolutions.antikythera.evaluator;
 
+import com.github.javaparser.ast.body.ConstructorDeclaration;
 import com.github.javaparser.ast.body.MethodDeclaration;
 import net.bytebuddy.implementation.bind.annotation.AllArguments;
 import net.bytebuddy.implementation.bind.annotation.Origin;
@@ -9,6 +10,7 @@ import sa.com.cloudsolutions.antikythera.evaluator.mock.MockingCall;
 import sa.com.cloudsolutions.antikythera.evaluator.mock.MockingRegistry;
 import sa.com.cloudsolutions.antikythera.parser.Callable;
 
+import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 
@@ -21,6 +23,57 @@ public class MethodInterceptor {
     }
     public MethodInterceptor(Class<?> clazz) {
         wrappedClass = clazz;
+    }
+
+    /**
+     * Intercept constructor calls with a ConstructorDeclaration.
+     * This is used when we have source code available.
+     */
+    @RuntimeType
+    public Object intercept(Constructor<?> constructor, Object[] args, ConstructorDeclaration constructorDecl) throws ReflectiveOperationException {
+        if (evaluator != null) {
+            // Push arguments onto stack in reverse order
+            for (int i = args.length - 1; i >= 0; i--) {
+                AntikytheraRunTime.push(new Variable(args[i]));
+            }
+
+            // Execute the constructor using source code evaluation
+            evaluator.executeConstructor(constructorDecl);
+
+            // For constructors, we need to return a new instance of the class
+            try {
+                return wrappedClass.getDeclaredConstructor().newInstance();
+            } catch (Exception e) {
+                // If we can't create an instance, return null
+                return null;
+            }
+        }
+        return intercept(constructor, args);
+    }
+
+
+    /**
+     * Intercepts constructor calls without a ConstructorDeclaration.
+     * This is used when we only have bytecode available.
+     */
+    @RuntimeType
+    public Object intercept(@Origin Constructor<?> constructor, @AllArguments Object[] args) throws ReflectiveOperationException {
+        if (wrappedClass != null) {
+            try {
+                Constructor<?> targetConstructor = wrappedClass.getDeclaredConstructor(constructor.getParameterTypes());
+                return targetConstructor.newInstance(args);
+            } catch (NoSuchMethodException e) {
+                // Constructor not found in wrapped class, fall through to default behavior
+            }
+        }
+
+        // For constructors, we need to return a new instance of the class
+        try {
+            return wrappedClass.getDeclaredConstructor().newInstance();
+        } catch (Exception e) {
+            // If we can't create an instance, return null
+            return null;
+        }
     }
 
     @RuntimeType
@@ -88,10 +141,10 @@ public class MethodInterceptor {
         this.wrappedClass = componentClass;
     }
 
-    public static class Interceptor {
+    public static class MethodDeclarationSupport {
         private final MethodDeclaration sourceMethod;
 
-        public Interceptor(MethodDeclaration sourceMethod) {
+        public MethodDeclarationSupport(MethodDeclaration sourceMethod) {
             this.sourceMethod = sourceMethod;
         }
 
@@ -102,6 +155,23 @@ public class MethodInterceptor {
             f.setAccessible(true);
             MethodInterceptor parent = (MethodInterceptor) f.get(instance);
             return parent.intercept(method, args, sourceMethod);
+        }
+    }
+
+    public static class ConstructorDeclarationSupport {
+        private final ConstructorDeclaration sourceConstructor;
+
+        public ConstructorDeclarationSupport(ConstructorDeclaration sourceConstructor) {
+            this.sourceConstructor = sourceConstructor;
+        }
+
+        @SuppressWarnings("java:S3011")
+        @RuntimeType
+        public Object intercept(@This Object instance, @Origin Constructor<?> constructor, @AllArguments Object[] args) throws ReflectiveOperationException {
+            Field f = instance.getClass().getDeclaredField(AKBuddy.CONSTRUCTOR_INTERCEPTOR);
+            f.setAccessible(true);
+            MethodInterceptor parent = (MethodInterceptor) f.get(instance);
+            return parent.intercept(constructor, args, sourceConstructor);
         }
     }
 }
